update to node 10 (new LTS) and move files to make it run and build properly
