+++ api-service/api-server/src/setup/custom_uploaders/upload_wiki_subtyped_rated.ts
@@ -20,59 +20,57 @@ import * as fs from 'fs';
 import * as yargs from 'yargs';
 
 // import * as wpconvlib from '@conversationai/wpconvlib';
-import * as crowdsourcedb from '../cs_db';
-import * as db_types from '../db_types';
+import * as crowdsourcedb from '../../cs_db';
+import * as db_types from '../../db_types';
 
-import {batchList} from './util'
+import {batchList} from '../util'
 
-    /*
+/*
 Usage:
 
-  node build/server/setup/upload_wiki_subtyped_rated.js \
-    --file="./tmp/subtypesdata/bad_test_x150.json" \
-    --question_group="wp_x2000" \
-    --question_type="test" \
-    --update_only
+node build/server/setup/upload_wiki_subtyped_rated.js \
+--file="./tmp/subtypesdata/bad_test_x150.json" \
+--question_group="wp_x2000" \
+--question_type="test" \
+--update_only
 
-  node build/server/setup/upload_wiki_subtyped_rated.js \
-    --file="./tmp/subtypesdata/unsure_test_x150.json" \
-    --question_group="wp_x2000" \
-    --question_type="test" \
-    --update_only
+node build/server/setup/upload_wiki_subtyped_rated.js \
+--file="./tmp/subtypesdata/unsure_test_x150.json" \
+--question_group="wp_x2000" \
+--question_type="test" \
+--update_only
 
-  node build/server/setup/upload_wiki_subtyped_rated.js \
-    --file="./tmp/subtypesdata/good_test_x150.json" \
-    --question_group="wp_x2000" \
-    --question_type="test" \
-    --update_only
+node build/server/setup/upload_wiki_subtyped_rated.js \
+--file="./tmp/subtypesdata/good_test_x150.json" \
+--question_group="wp_x2000" \
+--question_type="test" \
+--update_only
 
+node build/server/setup/upload_wiki_subtyped_rated.js \
+--file="./src/setup/subtypesdata/bad_train_x50.json" \
+--question_group="wp_x2000" \
+--question_type="training" \
+--update_only
 
+node build/server/setup/upload_wiki_subtyped_rated.js \
+--file="./src/setup/subtypesdata/unsure_train_x50.json" \
+--question_group="wp_x2000" \
+--question_type="training" \
+--update_only
 
-  node build/server/setup/upload_wiki_subtyped_rated.js \
-    --file="./src/setup/subtypesdata/bad_train_x50.json" \
-    --question_group="wp_x2000" \
-    --question_type="training" \
-    --update_only
+node build/server/setup/upload_wiki_subtyped_rated.js \
+--file="./src/setup/subtypesdata/good_train_x50.json" \
+--question_group="wp_x2000" \
+--question_type="training" \
+--update_only
 
-  node build/server/setup/upload_wiki_subtyped_rated.js \
-    --file="./src/setup/subtypesdata/unsure_train_x50.json" \
-    --question_group="wp_x2000" \
-    --question_type="training" \
-    --update_only
-
-  node build/server/setup/upload_wiki_subtyped_rated.js \
-    --file="./src/setup/subtypesdata/good_train_x50.json" \
-    --question_group="wp_x2000" \
-    --question_type="training" \
-    --update_only
-
-  node build/server/setup/upload_dataset.js \
-    --file="./src/testdata/wp_badish_x1000.json" \
-    --question_group="wp_x2000"
+node build/server/setup/upload_dataset.js \
+--file="./src/testdata/wp_badish_x1000.json" \
+--question_group="wp_x2000"
 */
 
-    // Command line arguments.
-    interface Params {file: string, question_type: string, question_group: string, gcloud_project_id: string, spanner_instance: string, spanner_db: string, update_only: boolean,};
+// Command line arguments.
+interface Params {file: string, question_type: string, question_group: string, gcloud_project_id: string, spanner_instance: string, spanner_db: string, update_only: boolean, };
 
 // type DataShape = wpconvlib.Conversation;
 interface DataShape {
@@ -90,8 +88,8 @@ interface DataShape {
 function makeScoreEnum(frac: number) {
   return {
     notatall: (frac < 0.3) ? 1 : (frac < 0.5 ? 0 : -1),
-        somewhat: (frac < 0.2) ? -0.5 : (frac < 0.8 ? 1 : -0.5),
-        very: (frac < 0.5) ? -1 : (frac < 0.7 ? 0 : 1),
+    somewhat: (frac < 0.2) ? -0.5 : (frac < 0.8 ? 1 : -0.5),
+    very: (frac < 0.5) ? -1 : (frac < 0.7 ? 0 : 1),
   }
 }
 
@@ -103,21 +101,21 @@ async function main(args: Params) {
   }
   if (!db_types.valid_question_type_regexp.test(args.question_type)) {
     console.error(
-        '--question_type must be a valid question type (training|test|toanswer), not: ' +
-        args.question_type);
+      '--question_type must be a valid question type (training|test|toanswer), not: ' +
+      args.question_type);
     return;
   }
   if (!db_types.valid_id_regexp.test(args.question_group)) {
     console.error(
-        '--question_type must be a valid question type (training|test|toanswer), not: ' +
-        args.question_type);
+      '--question_type must be a valid question type (training|test|toanswer), not: ' +
+      args.question_type);
     return;
   }
 
   let spannerClient = spanner({projectId: args.gcloud_project_id});
   let spannerInstance = spannerClient.instance(args.spanner_instance);
   let spannerDatabase =
-      spannerInstance.database(args.spanner_db, {keepAlive: 5});
+    spannerInstance.database(args.spanner_db, {keepAlive: 5});
   let db = new crowdsourcedb.CrowdsourceDB(spannerDatabase);
 
   let fileAsString = fs.readFileSync(args.file, 'utf8');
@@ -126,7 +124,7 @@ async function main(args: Params) {
   for (let i = 0; i < data.length; i++) {
     let id = data[i].rev_id;
     let question: db_types.Question = {id: id, text: data[i].comment_text} as
-        any;
+      any;
 
     let toxicity_scores = {enum: makeScoreEnum(parseFloat(data[i].frac_neg))};
     let readableAndInEnglishScores = {optional: true, enum: {yes: 1, no: -1}};
@@ -186,43 +184,43 @@ async function main(args: Params) {
 }
 
 let args =
-    yargs
-        .option(
-            'file',
-            {alias: 'f', describe: 'Path to JSON file of questions to upload'})
-        .option('question_type', {
-          alias: 't',
-          describe: 'Question type, one of: training | test | toanswer'
-        })
-        .option('question_group', {
-          alias: 'g',
-          describe: 'Question type, one of: training | test | toanswer'
-        })
-        .option(
-            'gcloud_project_id',
-            {alias: 'gcp', describe: 'Google Cloud Project Id'})
-        .option(
-            'spanner_instance', {alias: 'i', describe: 'Spanner instance name'})
-        .option('spanner_db', {alias: 's', describe: 'Spanner database name'})
-        .option('update_only', {
-          alias: 'u',
-          describe:
-              'If true, then do an update instead of an insert of the questons.'
-        })
-        .default('update_only', false)
-        .default('gcloud_project_id', 'wikidetox')
-        .default('spanner_instance', 'crowdsource')
-        .default('spanner_db', 'testdb')
-        .demandOption(
-            ['file', 'question_group', 'question_type'],
-            'Please provide at least --file and --question_group, and --question_type.')
-        .help()
-        .argv;
+  yargs
+    .option(
+      'file',
+      {alias: 'f', describe: 'Path to JSON file of questions to upload'})
+    .option('question_type', {
+      alias: 't',
+      describe: 'Question type, one of: training | test | toanswer'
+    })
+    .option('question_group', {
+      alias: 'g',
+      describe: 'Question type, one of: training | test | toanswer'
+    })
+    .option(
+      'gcloud_project_id',
+      {alias: 'gcp', describe: 'Google Cloud Project Id'})
+    .option(
+      'spanner_instance', {alias: 'i', describe: 'Spanner instance name'})
+    .option('spanner_db', {alias: 's', describe: 'Spanner database name'})
+    .option('update_only', {
+      alias: 'u',
+      describe:
+        'If true, then do an update instead of an insert of the questons.'
+    })
+    .default('update_only', false)
+    .default('gcloud_project_id', 'wikidetox')
+    .default('spanner_instance', 'crowdsource')
+    .default('spanner_db', 'testdb')
+    .demandOption(
+      ['file', 'question_group', 'question_type'],
+      'Please provide at least --file and --question_group, and --question_type.')
+    .help()
+    .argv;
 
 main(args as any as Params)
-    .then(() => {
-      console.log('Success!');
-    })
-    .catch(e => {
-      console.error('Failed: ', e);
-    });
+  .then(() => {
+    console.log('Success!');
+  })
+  .catch(e => {
+    console.error('Failed: ', e);
+  });
